refactor(orchestration): remove 2 empty orchestrators — RealLLM shim + LogiqueComplexe stub (#885)
Remove deprecated pre-Registry orchestrators that are literally empty:

1. RealLLMOrchestrator (124 LOC shim) — NotImplementedError since 2026-03
   - Dataclasses migrated to orchestration/models.py
   - RealConversationLogger alias replaced by direct ConversationLogger import
   - 5 callers cleaned (service_manager, main_orchestrator, unified_pipeline,
     unified_text_analysis, pipelines/__init__)

2. LogiqueComplexeOrchestrator (108 LOC stub) — hardcoded mock since creation
   - run_einstein_puzzle() returned hardcoded 'Allemand possède le poisson'
   - 4 callers cleaned (orchestration/__init__, main_orchestrator,
     logic_orchestrator wrapper, pipelines/__init__)

3. RealLLMOrchestratorWrapper (pipelines/ specialized shim) — also deleted

4. test_real_llm_orchestrator.py — tests targeted deleted dataclasses only

Total: -4 files, ~340 LOC removed, +1 models.py (dataclass preservation)

## Files removed and why

| File | Type | Last useful | Reason |
|------|------|-------------|--------|
| real_llm_orchestrator.py | shim | 2026-03 | NotImplementedError shim, superseded by UnifiedPipeline |
| logique_complexe_orchestrator.py | stub | 2025-06 | Hardcoded mock, superseded by FOL/Tweety Registry |
| pipelines/.../real_llm_orchestrator.py | wrapper | 2026-05 | Thin wrapper over dead shim |
| test_real_llm_orchestrator.py | test | 2026-05 | Tests dataclass-only, dataclasses migrated |

Origin: Sprint 13 (2026-03) shimming strategy. 3 months later, Registry stable, callers migrated.

281 tests pass, 0 new failures. 1 pre-existing failure unrelated (spectacular count).
