--- argumentation_analysis/orchestration/logique_complexe_orchestrator.py
@@ -1,108 +1,0 @@
-# argumentation_analysis/orchestration/logique_complexe_orchestrator.py
-
-import logging
-import warnings
-import asyncio
-from typing import Optional, List, Dict, Any
-from semantic_kernel import Kernel
-
-# PURGE PHASE 3A: Utilisation des définitions minimales de cluedo_extended_orchestrator
-# pour Agent, SelectionStrategy, etc. car semantic_kernel.agents n'est pas disponible.
-# Note: AgentGroupChat et ChatCompletionAgent ne sont pas directement remplacés ici,
-# leur usage devra être adapté ou ces classes devront être définies localement si nécessaires.
-# Pour l'instant, on commente les imports directs qui échoueraient.
-# from semantic_kernel.agents import AgentGroupChat, ChatCompletionAgent # N'existe pas dans SK 0.9.6b1
-
-# Import des définitions de base depuis l'orchestrateur principal
-from argumentation_analysis.orchestration.cluedo_components.strategies import (
-    CyclicSelectionStrategy,
-)
-from .base import Agent, SelectionStrategy, TerminationStrategy
-
-# Si AgentGroupChat ou ChatCompletionAgent sont réellement utilisés, il faudra les définir ici
-# ou adapter le code pour utiliser des mécanismes d'orchestration plus simples.
-# Les définitions locales de SequentialSelectionStrategy, ChatCompletionAgent et AgentGroupChat ont été supprimées.
-# Il faudra s'assurer que les usages restants sont compatibles avec les classes de semantic_kernel 1.32.2
-# ou avec les définitions de cluedo_extended_orchestrator.
-
-# Tentative d'import des classes AgentGroupChat et ChatCompletionAgent depuis semantic_kernel.agents
-# Si cela échoue, le code qui les utilise devra être adapté.
-# try:
-#     from semantic_kernel.agents import AgentGroupChat as SKAgentGroupChat
-#     from semantic_kernel.agents import ChatCompletionAgent as SKChatCompletionAgent
-#     # Si l'import réussit, on pourrait les utiliser. Sinon, il faudra une autre solution.
-# except ImportError:
-#     # Fallback: si les imports directs échouent, on loggue un avertissement.
-#     # Le code plus bas qui utilise AgentGroupChat ou ChatCompletionAgent pourrait planter
-#     # ou devra être adapté pour utiliser GroupChatOrchestration ou Agent de cluedo_extended_orchestrator.
-#     logging.warning("Impossible d'importer AgentGroupChat ou ChatCompletionAgent depuis semantic_kernel.agents. "
-#                     "Les fonctionnalités dépendantes pourraient être affectées.")
-#     # On définit des placeholders pour éviter des NameError immédiats si le code n'est pas entièrement purgé
-#     # de leurs références, mais cela ne les rendra pas fonctionnels.
-#     class SKAgentGroupChat: pass
-#     class SKChatCompletionAgent(Agent): pass # Hérite de notre Agent de base pour un minimum de structure
-
-
-# Le code suivant qui instancie AgentGroupChat devra être vérifié.
-# S'il utilisait la version locale, il doit maintenant être compatible avec SKAgentGroupChat
-# ou une alternative.
-
-# Exemple d'adaptation (si AgentGroupChat local était utilisé) :
-# La classe LogiqueComplexeOrchestrator pourrait avoir besoin d'être révisée
-# pour utiliser GroupChatOrchestration de cluedo_extended_orchestrator,
-# ou pour que SKAgentGroupChat soit compatible.
-
-# Pour l'instant, on se concentre sur la suppression des définitions locales.
-# La logique d'instanciation de AgentGroupChat dans ce fichier est :
-# class AgentGroupChat: (supprimée)
-#   def __init__(... selection_strategy: Optional[SelectionStrategy] = None ...):
-#       self.selection_strategy = selection_strategy or SequentialSelectionStrategy(self.agents) (SequentialSelectionStrategy locale supprimée)
-
-# Si une classe AgentGroupChat est toujours nécessaire ici, elle devrait être SKAgentGroupChat.
-# Et sa `selection_strategy` devrait être compatible.
-# `CyclicSelectionStrategy` est importée depuis cluedo_extended_orchestrator.
-
-# Si ce fichier définit une classe qui hérite ou utilise AgentGroupChat,
-# cette partie devra être adaptée.
-# Par exemple, si une classe OrchestrateurLogiqueComplexe existe et fait :
-# self.chat = AgentGroupChat(agents=..., selection_strategy=CyclicSelectionStrategy(...))
-# alors AgentGroupChat doit être SKAgentGroupChat et compatible.
-
-# Pour l'instant, je supprime les définitions locales.
-# La correction complète de l'utilisation de AgentGroupChat et ChatCompletionAgent
-# dans ce fichier dépendra de leur usage réel plus bas, que je n'ai pas encore vu.
-# Je vais supposer pour l'instant que le code plus bas sera adapté ou n'utilise pas ces versions locales.
-
-# La classe LogiqueComplexeOrchestrator elle-même n'est pas dans cet extrait.
-# Je vais me concentrer sur la suppression des définitions de classes fallbacks.
-# L'initialisation de `self.selection_strategy` dans la classe AgentGroupChat locale
-# utilisait `SequentialSelectionStrategy(self.agents)`.
-# Si `SKAgentGroupChat` est utilisé, il faudra voir comment il gère la stratégie.
-# `CyclicSelectionStrategy` est maintenant importée et pourrait être passée à `SKAgentGroupChat`
-# si son constructeur accepte un `selection_strategy`.
-
-
-# Définition minimale pour LogiqueComplexeOrchestrator
-class LogiqueComplexeOrchestrator:
-    def __init__(self, kernel: Kernel = None, **kwargs):
-        warnings.warn(
-            "`LogiqueComplexeOrchestrator` is deprecated and contains minimal logic. "
-            "It is maintained for backward compatibility only. "
-            "Please use the new agent group chat architecture for new implementations.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        self._logger = logging.getLogger(self.__class__.__name__)
-        self.kernel = kernel
-        self._logger.info(
-            "LogiqueComplexeOrchestrator initialisé (définition minimale)."
-        )
-
-    async def run_einstein_puzzle(self, puzzle_data: Dict[str, Any]) -> Dict[str, Any]:
-        self._logger.info("Exécution du puzzle Einstein (simulation minimale).")
-        # Simulation d'une exécution
-        await asyncio.sleep(0.01)
-        return {
-            "solution": "L'Allemand possède le poisson (simulation)",
-            "success": True,
-        }
