+++ argumentation_analysis/pipelines/orchestration/orchestrators/specialized/logic_orchestrator.py
@@ -1,8 +1,11 @@
 #!/usr/bin/env python3
-# -*- coding: utf-8 -*-
+# -*- coding: utf-8-
 
 """
 Specialized Complex Logic Orchestrator Core Module.
+
+LogiqueComplexeOrchestrator removed (#885) — superseded by FOL/Tweety via Registry.
+This wrapper is retained as a placeholder for future logic orchestration integration.
 """
 
 import logging
@@ -10,36 +13,27 @@ from typing import Dict, Any
 
 logger = logging.getLogger(__name__)
 
-try:
-    from argumentation_analysis.orchestration.logique_complexe_orchestrator import (
-        LogiqueComplexeOrchestrator,
-    )
-except ImportError as e:
-    logger.warning(f"LogiqueComplexeOrchestrator not available: {e}")
-    LogiqueComplexeOrchestrator = None
-
 
 class LogicOrchestratorWrapper:
+    """Placeholder for logic orchestration.
+
+    The original LogiqueComplexeOrchestrator was a deprecated stub returning
+    hardcoded mock data. It has been removed. Future logic orchestration
+    should use FOLLogicAgent + TweetyLogicPlugin via CapabilityRegistry.
+    """
+
     def __init__(self):
-        if not LogiqueComplexeOrchestrator:
-            raise ImportError("LogiqueComplexeOrchestrator is not available.")
-        self.orchestrator = LogiqueComplexeOrchestrator()
-        logger.info("[LOGIC_COMPLEX] LogiqueComplexeOrchestrator initialized.")
+        logger.info(
+            "[LOGIC_COMPLEX] LogicOrchestratorWrapper initialized (stub — "
+            "LogiqueComplexeOrchestrator removed #885)."
+        )
 
     async def analyze(self, text: str) -> Dict[str, Any]:
-        """
-        Runs a complex logic analysis.
-        """
-        logger.info("[LOGIC_COMPLEX] Running complex logic analysis...")
-        if not hasattr(self.orchestrator, "analyze_complex_logic"):
-            return {
-                "status": "error",
-                "error": "Method 'analyze_complex_logic' not found in orchestrator.",
-            }
-
-        try:
-            results = await self.orchestrator.analyze_complex_logic(text)
-            return {"status": "completed", "logic_analysis": results}
-        except Exception as e:
-            logger.error(f"Error during complex logic analysis: {e}", exc_info=True)
-            return {"status": "error", "error": str(e)}
+        """Returns a deprecation notice — use Registry-based logic agents instead."""
+        return {
+            "status": "deprecated",
+            "message": (
+                "LogiqueComplexeOrchestrator has been removed (#885). "
+                "Use FOLLogicAgent or TweetyLogicPlugin via CapabilityRegistry."
+            ),
+        }
